Next step -
Ready to integrate logic into the board itself. The board needs more work before it is ready to have logic implemented. Ideally, the logic will be implemented in a reducer so the main function of the Board component will be to elegantly display the big thoughts of what needs to happen while the minutia of what's actually happening occurs behind the scenes.